requirements.txt
astroid
autopep8
certifi
chardet
defusedxml
Django==2.2.13
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn
idna
isort
lazy-object-proxy
mccabe
oauthlib
pep8
pycodestyle
pylint
python3-openid
pytz
requests
requests-oauthlib
six
sqlparse
typed-ast
urllib3
wrapt
django-countries
pillow